Improve sync point accuracy

+++ src/insound/SyncPointInfo.h
@@ -7,6 +7,6 @@ namespace Insound
     struct SyncPointInfo
     {
         std::string name;
-        unsigned offset;
+        double offset;
     };
 }
